+++ ft_printf.h
@@ -7,10 +7,9 @@ int ft_printf(const char *str, ...);
 int ft_putchar_fd(char c, int fd);
 int ft_putstr_fd(char *s, int fd);
 int ft_putnbr_fd(int n, int fd);
-int ft_putnbr_pos_fd(unsigned int n, int fd);
+int ft_putnbr_pos_fd(unsigned long n, int fd);
 int ft_putnbr_hex_fd(unsigned long n, int fd, unsigned int cap);
-// int	ft_putnbr_hex_fd(unsigned long n, int fd, int cap);
 int ft_format_specifiers(char fs, va_list args);
-int ft_putadd_fd(int n, int fd);
+int ft_putadd_fd(unsigned long n, int fd);
 
 #endif
